Collapsed various header and footer files.
Created a single header.php file and a single footer.php file. Moved content
from files such as knowvim-navigation.php directly into header.php.  The
file is larger but I think it's now simpler to understand what is being
included where. Eventually I need to change this naive approach to
the server-side rendering of pages, but it works for now and the site is
simple enough that I can still maintain this system as-is.