Added crude front controller.
index.php is a home-brewed front controller that just inserts a header and
footer around the main VimWindow "buffer" contents. This is crude, but at
least a bit better than having to put require("header.php") and
require("footer.php") statements at the top and bottom of every new page in
the site.