Undid the front controller stuff.
Maybe adding require("header.php") and require("footer.php") isn't so bad
after all. I realized the Rewrite rules are a little bit tricky since if you
ask for a resource like /explain.php and that resource actually exists as a
file, then it will serve that up rather than applying the Rewrite rule
first. I'm sure there is a way to stop this but I'm just going to go back
to the old way for now.