Created Buffer model and EditorView view.
The Buffer model is wired to fetch text files from /files/ on the server and
split the list of lines into an array. The EditorView has a render()
function which iterates through each line in this array (via
view.buffer.get('lines')) and wraps the contents in appropriate HTML markup
tags. The resulting markup is inserted into the EditorView's el element.