Very basic h, j, k, l motion handling.
Stubbed out cases for the 'h', 'j', 'k', and 'l' motions in
getMotionResult. The function returns an object with attributes for the
starting positions, the ending positions, the type of motion (linewise or
characterwise), and whether or not the motion is inclusive. Certain edge
cases not yet accounted for (e.g., adjusting the endCol when making a 'j' or
'k' movement).