Implemented better Logger.
The Logger constructor takes an moduleName string in its constructor and
searches for this string in the window's query string. If the moduleName
string appears somewhere after the "log" query parameter then the message is
sent to the console. Otherwise, nothing happens.

The module names are as follows:

?log=vim        : Messages from Vim.js
?log=normal     : Messages from NormalHandler.js
?log=editor     : Messages from Editor.js
?log=parser     : Messages from Parser.js
?log=tokenizer  : Messages from Tokenizer.js
?log=operations : Messages from Operations.js

Supports multiple module names in the query string. They should be delimited
with semicolons. Example:

?log=tokenizer;parser : Messages from both Parser.js and Tokenizer.js