Implemented mark operations.
Operations.getOperationResult now expects a "vim" property in the arguments
object that points to the Vim model. The reference is used in
MarkOperations.getMarkOperationResult when setting the "marks" property of
the Vim model.

The "marks" property is an object whose keys are mark names and whose values
are objects containing the "row" and "col" of the mark. Note that the
buffer associated with the mark is not included in this information. There
is also no logic to handle the shifting of mark positions when lines are
added or deleted from the buffer.